Informative error where type cannot be converted to Open API type.
When a user-defined generic type is resolved to be a parameter, we are unable to determine an Open API type.

In this case, it is highly likely that the function parameter is for a dependency that hasn't been provided.

This change catches the key error and raises `ImproperlyConfiguredException` with a more user-friendly message.
